Moved config files Added WIP ClawF.kt
