+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/telop/ClawTest.kt
@@ -6,8 +6,8 @@ import com.qualcomm.robotcore.eventloop.opmode.TeleOp
 import org.firstinspires.ftc.teamcode.modules.drive.HDrive
 import org.firstinspires.ftc.teamcode.modules.drive.SparkfunImuLocalizer
 import org.firstinspires.ftc.teamcode.modules.hardware.GamepadEx
-import org.firstinspires.ftc.teamcode.modules.robot.Claw
-import org.firstinspires.ftc.teamcode.modules.robot.Slide
+import org.firstinspires.ftc.teamcode.opmode.config.Claw
+import org.firstinspires.ftc.teamcode.opmode.config.Slide
 import org.firstinspires.ftc.teamcode.opmode.config.HDriveConfig
 
 @TeleOp
