+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/telop/ESampleTest.kt
@@ -6,8 +6,8 @@ import com.qualcomm.robotcore.eventloop.opmode.TeleOp
 import org.firstinspires.ftc.teamcode.modules.drive.HDrive
 import org.firstinspires.ftc.teamcode.modules.drive.SparkfunImuLocalizer
 import org.firstinspires.ftc.teamcode.modules.hardware.GamepadEx
-import org.firstinspires.ftc.teamcode.modules.robot.Arm
-import org.firstinspires.ftc.teamcode.modules.robot.Spin
+import org.firstinspires.ftc.teamcode.opmode.config.Arm
+import org.firstinspires.ftc.teamcode.opmode.config.Spin
 import org.firstinspires.ftc.teamcode.opmode.config.HDriveConfig
 
 @TeleOp
